pyproject.toml
[project]
name = "job-market-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "apache-airflow>=3.1.8",
    "appnope==0.1.4",
    "asttokens==3.0.1",
    "async-generator==1.10",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "comm==0.2.3",
    "cronitor>=4.9.0",
    "databricks-sdk>=0.47.0",
    "dbt-core>=1.11.5",
    "dbt-databricks>=1.10.9",
    "dbt-postgres>=1.10.0",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "executing==2.2.1",
    "h11==0.16.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "ipykernel==7.1.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "markdownify==0.13.1",
    "matplotlib-inline==0.2.1",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==1.26.3",
    "outcome==1.3.0.post0",
    "packaging==26.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "platformdirs==4.5.1",
    "prompt-toolkit==3.0.52",
    "psutil==7.2.2",
    "psycopg2>=2.9.11",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pysocks==1.7.1",
    "pyspark>=4.1.1",
    "python-crontab>=3.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.2",
    "python-jobspy==1.1.82",
    "pytz==2025.2",
    "pyzmq==27.1.0",
    "regex==2024.11.6",
    "requests==2.32.5",
    "schedule>=1.2.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.8.3",
    "stack-data==0.6.3",
    "tls-client==1.0.1",
    "tornado==6.5.4",
    "traitlets==5.14.3",
    "trio==0.32.0",
    "trio-typing==0.10.0",
    "trio-websocket==0.12.2",
    "types-certifi==2021.10.8.3",
    "types-urllib3==1.26.25.14",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "wcwidth==0.5.0",
    "websocket-client==1.9.0",
    "wsproto==1.3.2",
    "zipp==3.23.0",
]
